DOCTECH-BE/DOCTECH-TEXT_EXTRATION-SERVICE/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
Deprecated==1.2.14
fastapi==0.112.0
filetype==1.2.0
h11==0.14.0
idna==3.7
importlib_resources==6.4.0
limits==3.13.0
packaging==24.1
pdf2image==1.17.0
pdfminer.six           # Removed version constraint
pdfplumber             # Removed version constraint
pillow==10.4.0
psycopg2==2.9.9
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pypdfium2==4.30.0
pytesseract==0.3.10
python-dotenv==1.0.1
setuptools==71.1.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.5
wrapt==1.16.0
textract==1.6.4
python-docx             # Added to the appropriate place
python-multipart